requirements.txt
Flask==2.0.1
pyttsx3==2.90
SpeechRecognition==3.8.1
plyer==2.0.0
pyautogui==0.9.53
wikipedia-api==0.5.4
pywhatkit==5.3
